Move Verbosity to IO/Logging
The Informer library is quite low-level, and it is difficult enough to
link as it is (because of the InfoAtLink mechanism).
The logging-related functionality better has its own library.
