+++ src/IO/Logging/CMakeLists.txt
@@ -1,37 +1,30 @@
 # Distributed under the MIT License.
 # See LICENSE.txt for details.
 
-set(LIBRARY LinearSolver)
+set(LIBRARY Logging)
 
 add_spectre_library(${LIBRARY})
 
 spectre_target_sources(
   ${LIBRARY}
   PRIVATE
-  Gmres.cpp
-  Lapack.cpp
+  Verbosity.cpp
   )
 
 spectre_target_headers(
   ${LIBRARY}
   INCLUDE_DIRECTORY ${CMAKE_SOURCE_DIR}/src
   HEADERS
-  ExplicitInverse.hpp
-  Gmres.hpp
-  InnerProduct.hpp
-  Lapack.hpp
-  LinearSolver.hpp
+  Tags.hpp
+  Verbosity.hpp
   )
 
 target_link_libraries(
   ${LIBRARY}
   PUBLIC
-  Blas
-  Convergence
-  DataStructures
-  Logging
   Options
-  Parallel
   PRIVATE
-  Lapack
+  ErrorHandling
+  INTERFACE
+  DataStructures
   )
